Refactored Auth
removed component Auth and login and register js

added Auth utility that handles the exact same thing
AuthLogin did

AuthRegister was not used anywhere within the App
so removed
