+++ client/src/utils/Auth.js
